// src/api/generated/models/storeDevProfileRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Brew
 * OpenAPI spec version: 0.0.1
 */
import type { StoreDevProfileRequestOpenToRelocation } from './storeDevProfileRequestOpenToRelocation';
import type { StoreDevProfileRequestOpenToWork } from './storeDevProfileRequestOpenToWork';
import type { SeniorityLevelEnum } from './seniorityLevelEnum';
import type { DevSpecialtyEnum } from './devSpecialtyEnum';

export interface StoreDevProfileRequest {
  cpf: string;
  /**
   * @minLength 3
   * @maxLength 255
   */
  name: string;
  bio: string;
  phone: string;
  open_to_relocation?: StoreDevProfileRequestOpenToRelocation;
  open_to_work?: StoreDevProfileRequestOpenToWork;
  birthdate: string;
  seniority_level: SeniorityLevelEnum;
  specialty: DevSpecialtyEnum;
}